+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `create-config` would fail with `--path` specified.
+
 ## [2.1.1] - 2022-05-29
 
 ### Fixed